Log output on failed publish
We were in a situation with the backstage instance where the publish
action was running and succeeding, but the packages were nowhere to be
found. It turned out that there was a failure because [the current
working directory of the package was incorrect][1]. However,
diagnosing this was extremely difficult because it did not provide any
indication as to what went wrong.

This prints out details about the command including what its output is
in order to help debug in the future. I would argue that we should
eventually make the action fail completely if any of the packages are
not properly published, but at the very least we want to figure out why.

[1]: https://github.com/thefrontside/actions/pull/77