Synchronize to NPM
The publication scripts weren't quite complete, and so could not be
used in repos generally. They did not:

1. return found packages to main script
2. compute package paths correctly
3. filter out packages from nested node_modules/

This adds those features as well as a DRY_RUN input that lets you run
the action without actually attempting to publish to NPM. In addittion
so streaming logs are added that are foldable to help debugging.